colony display
